refactor:为FeedItem添加layoutMode支持以优化布局切换性能

+++ app/src/main/java/com/example/bytedancehomework/tracker/ExposureTracker.java
@@ -140,7 +140,7 @@ public class ExposureTracker {
     }
 
 
-    private int getScreenHeight(View view) {
+    private float getScreenHeight(View view) {
         DisplayMetrics metrics = view.getResources().getDisplayMetrics();
         return metrics.heightPixels;
     }
